Move fitting functionality into plotting model
RE #34587
